put appstream interaction behind mutex
I am seeing an obscene amount of crashes that most likely come from
libappstream(qt) not being thread safe (i.e. all the backtraces I got
recently were with 2 concurrent matches going on).
To prevent crashery, put the appstream database interaction behind a mutex
lock, isolate the entire thing into a function for convenience (Components
are qsharedata, so that's fine). This is however still a bit naughty as we
only read from the Component objects after the lock, which may or may not
crash in the future as Components right now look fully complete, nothing's
preventing them from being resolved on-demand (i.e. also hit the database)
in the future though, at which point they would be unguarded critical
paths.
Also this means KServiceTypeTrader is not guarded as I am not sure it needs
a mutex guard at all or if it is thread safe. Might be good to check up
on this any if it isn't save simply cover the entire match function in
a mutexlocker.

NB: only one runner is initalized, but it's matching is run concurrently

CCMAIL: aleixpol@kde.org